pyproject.toml
[project]
name = "anycost-generator"
version = "0.1.0"
description = "AnyCost Adaptor Generator - generates customized CloudZero AnyCost Stream adaptors"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "InquirerPy>=0.3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
anycost-generator = "anycost_generator.cli.main:main"

[tool.setuptools.packages.find]
include = ["anycost_generator*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
